[lib/carpet] Reverted local lib change

+++ lib/carpet/src/utils/vector/crpt_vec_magnitude.c
@@ -30,7 +30,7 @@ double crpt_vec3_magnitude(vec3_t vec)
 */
 double crpt_vec2_magnitude(vec2_t vec)
 {
-    return sqrtf(
+    return sqrt(
         vec.x * vec.x +
         vec.y * vec.y
     );
